Рефакторинг скриптов для работы со схемой БД.

+++ scripts/dumpdb.sh
@@ -1,15 +1,17 @@
-#!/bin/bash
+#!/bin/sh
 
-if [[ $(dirname $0) == "." ]]; then
-	printf "%s\n" "Call script from root repo." >&2 && exit 1
+if [ ! -f schema.sql ]; then
+	echo "Can't find schema.sql. You running script from root repo?" >&2
+	exit 1
 fi
 
-if [[ $# -ne 2 ]]; then
-	printf "%s\n" "Usage: $0 <db_name> <db_user>" >&2 && exit 2
+if [ $# -ne 2 ]; then
+	echo "Usage: $0 <db_name> <db_user>" >&2
+	exit 2
 fi
 
 echo "Dumping database..."
-pg_dump -v -U $2 -d $1 -sxO > schematmp.sql || ( rm schematmp.sql && exit 3 )
+pg_dump --username="$2" --dbname="$1" --schema-only --no-privileges --no-owner --clean --if-exists > schematmp.sql || { rm schematmp.sql; exit 3; }
 echo "Replacing original schema..."
 mv schematmp.sql schema.sql
 echo "Dumping completed."
